Add tabs.  Componentize future appointments.
