+++ src/applications/vaos/sass/vaos.scss
@@ -1,6 +1,35 @@
 @import "~@department-of-veterans-affairs/formation/sass/shared-variables";
 @import "../../../platform/forms/sass/m-schemaform";
 
+.vaos-appts__tabs {
+
+  ul.react-tabs__tab-list {
+    padding: 0;
+    border-bottom: 1px solid $color-gray-light;
+  }
+
+  .vaos-appts__tab {
+    border-color: $color-gray-light;
+    text-align: center;
+    background: $color-gray-light-alt;
+    border: solid 1px $color-gray-light;
+    border-radius: 0 !important;
+    margin: 0;
+    display: inline-block;
+    bottom: -1px;
+    position: relative;
+    list-style: none;
+    padding: 6px 12px;
+    cursor: pointer;
+
+    &.react-tabs__tab--selected {
+      border-top: solid 3px $color-primary;
+      background: $color-white;
+      border-bottom: solid 1px $color-white;
+    }
+  }
+}
+
 .vaos-option-list__icon {
   display: inline-block;
   font-size: 19px;
